Integrated current weather model onto home screen

+++ lib/models/current_weather_model.dart
@@ -8,7 +8,7 @@ LocationService cLocationService = LocationService();
 class CurrentWeatherModel {
    int? windSpeed;
    int? humidity;
-   int? currentTemperature;
+   double? currentTemperature;
 
   CurrentWeatherModel({
      this.windSpeed,
